+++ src/Components/Inputs/Input.tsx
@@ -10,7 +10,7 @@ import React, {
 import classnames from 'classnames'
 
 // Components
-import {Icon} from '../Icon/Icon'
+import {Icon} from '../Icon/Base/Icon'
 import {StatusIndicator} from './StatusIndicator'
 
 // Styles