+++ src/Components/Inputs/Toggle.tsx
@@ -9,7 +9,7 @@ import React, {
 import classnames from 'classnames'
 
 // Components
-import {Icon} from '../Icon/Icon'
+import {Icon} from '../Icon/Base/Icon'
 
 // Styles
 import './Toggle.scss'