+++ src/Components/Panel/index.tsx
@@ -4,6 +4,7 @@ import React, {Component} from 'react'
 // Components
 import {PanelRoot, PanelProps} from './Family/Panel'
 import {PanelHeader} from './Family/PanelHeader'
+import {PanelSymbolHeader} from './Composed/PanelSymbolHeader'
 import {PanelBody} from './Family/PanelBody'
 import {PanelFooter} from './Family/PanelFooter'
 
@@ -12,6 +13,7 @@ export class Panel extends Component<PanelProps> {
 
   public static Panel = PanelRoot
   public static Header = PanelHeader
+  public static SymbolHeader = PanelSymbolHeader
   public static Body = PanelBody
   public static Footer = PanelFooter
 
@@ -22,5 +24,6 @@ export class Panel extends Component<PanelProps> {
 
 export {PanelProps, PanelRef} from './Family/Panel'
 export * from './Family/PanelHeader'
+export * from './Composed/PanelSymbolHeader'
 export * from './Family/PanelBody'
 export * from './Family/PanelFooter'